Refactor solution architecture to allow for two separate projects
- Issue: original project contained solution and back-end project in the same folder. In that folder the UI project was created
- Solution: made a new tempated-based solution (.NET + ReactTS) and applied the previous source code to it. The nw solution has front-end and back-end as separate projects under one single solution.
