+++ TaskManagementSystem.Server/API/Controllers/TodoTaskController.cs
@@ -1,0 +1,86 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using System.Globalization;
+using TaskManagementSystem.Server.API.Models;
+using TaskManagementSystem.Server.Data;
+using TaskManagementSystem.Server.Models;
+
+namespace TaskManagementSystem.Server.API.Controllers;
+
+[ApiController]
+[Route("api/tasks")]
+public class TodoTaskControllers(IToDoTaskRepository repository) : ControllerBase
+{
+    [HttpGet]
+    public ActionResult<IEnumerable<ToDoTask>> GetAll(string? title = null, string? dueDate = null, string? priority = null)
+    {
+        var tasks = repository.GetAll();
+
+        if (title is not null)
+            tasks = tasks.Where(t => t.Title.Contains(title));
+
+        if (dueDate is not null && DateTime.TryParse(dueDate,
+                new DateTimeFormatInfo() { FullDateTimePattern = "yyyy-MM-dd" },
+                out var dueDateParsed))
+            tasks = tasks.Where(t => t.DueDate == dueDateParsed);
+
+        if (priority is not null)
+            tasks = tasks.Where(t => t.Priority == priority);
+
+        return Ok(tasks);
+    }
+
+    [HttpPost]
+    public ActionResult Add([FromBody] TaskDTO task)
+    {
+        if (!DateTime.TryParse(
+            task.DueDate,
+            new DateTimeFormatInfo() { FullDateTimePattern = "yyyy-MM-dd" },
+            out var dueDateParsed) || string.IsNullOrWhiteSpace(task.Title))
+            return BadRequest();
+
+        repository.Add((new ToDoTask()
+        {
+            Description = task.Description,
+            DueDate = dueDateParsed,
+            Priority = task.Priority,
+            Title = task.Title
+        }));
+
+        return Ok();
+    }
+
+    [HttpPut("{id}")]
+    public ActionResult Update(int id, [FromBody] TaskDTO task)
+    {
+        var retrievedTask = repository.GetById(id);
+
+        if (retrievedTask is null)
+            return NoContent();
+
+        if (!DateTime.TryParse(
+                task.DueDate,
+                new DateTimeFormatInfo() { FullDateTimePattern = "yyyy-MM-dd" },
+                out var dueDateParsed))
+            return BadRequest();
+
+        retrievedTask.Title = task.Title;
+        retrievedTask.Description = task.Description;
+        retrievedTask.DueDate = dueDateParsed;
+        retrievedTask.Priority = task.Priority;
+
+        repository.Update(retrievedTask);
+
+        return Ok(retrievedTask);
+    }
+
+    [HttpDelete("{id}")]
+    public ActionResult Delete(int id)
+    {
+        var task = repository.GetById(id);
+        if (task is null)
+            return NotFound();
+
+        repository.Delete(id);
+        return Ok();
+    }
+}
