+++ TaskManagementSystem.Server/Data/IToDoTaskRepository.cs
@@ -1,0 +1,12 @@
+﻿using TaskManagementSystem.Server.Models;
+
+namespace TaskManagementSystem.Server.Data;
+
+public interface IToDoTaskRepository
+{
+    IEnumerable<ToDoTask> GetAll();
+    ToDoTask? GetById(int id);
+    void Add(ToDoTask task);
+    void Update(ToDoTask task);
+    void Delete(int id);
+}
