+++ TaskManagementSystem.Server/TaskManagementSystem.Server.http
@@ -1,0 +1,6 @@
+@TaskManagementSystem.Server_HostAddress = http://localhost:5113
+
+GET {{TaskManagementSystem.Server_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
