+++ TaskManagementSystem.slnx
@@ -1,0 +1,7 @@
+<Solution>
+  <Project Path="taskmanagementsystem.client/taskmanagementsystem.client.esproj">
+    <Build />
+    <Deploy Solution="Release|*" />
+  </Project>
+  <Project Path="TaskManagementSystem.Server/TaskManagementSystem.Server.csproj" />
+</Solution>
