--- TaskManager.Api.http
@@ -1,6 +1,0 @@
-@TaskManager.Api_HostAddress = http://localhost:5118
-
-GET {{TaskManager.Api_HostAddress}}/api/tasks
-Accept: application/json
-
-###
