--- TaskManager.Api.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.14.36221.1 d17.14
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TaskManager.Api", "TaskManager.Api.csproj", "{094DDE61-C7B1-57AB-B464-6EFC1C33B56C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{094DDE61-C7B1-57AB-B464-6EFC1C33B56C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{094DDE61-C7B1-57AB-B464-6EFC1C33B56C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{094DDE61-C7B1-57AB-B464-6EFC1C33B56C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{094DDE61-C7B1-57AB-B464-6EFC1C33B56C}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {96FD0C25-A0C0-4A3B-9CFA-897C097CAAC5}
-	EndGlobalSection
-EndGlobal
