--- task-manager-ui/vite.config.ts
@@ -1,16 +1,0 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-
-// https://vite.dev/config/
-export default defineConfig({
-  plugins: [react()],
-  server: {
-    proxy: {
-      '/api': {
-        target: 'http://localhost:5118',
-        changeOrigin: true,
-        secure: false
-      }
-    }
-  }
-})
