+++ taskmanagementsystem.client/CHANGELOG.md
@@ -1,0 +1,13 @@
+This file explains how Visual Studio created the project.
+
+The following tools were used to generate this project:
+- create-vite
+
+The following steps were used to generate this project:
+- Create react project with create-vite: `npm init --yes vite@latest taskmanagementsystem.client -- --template=react-ts`.
+- Create project file (`taskmanagementsystem.client.esproj`).
+- Create `launch.json` to enable debugging.
+- Add project to solution.
+- Update proxy endpoint to be the backend server endpoint.
+- Add project to the startup projects list.
+- Write this file.
