+++ taskmanagementsystem.client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "task-manager-ui",
+  "name": "taskmanagementsystem.client",
   "private": true,
   "version": "0.0.0",
   "type": "module",
