+++ taskmanagementsystem.client/src/App.tsx
@@ -1,6 +1,6 @@
-import {useEffect, useState} from "react";
-import {getTasks} from "./services/taskService";
-import type {ToDoTask} from "./types/ToDoTask.ts";
+import { useEffect, useState } from "react";
+import { getTasks } from "./services/taskService";
+import type { ToDoTask } from "./types/ToDoTask.ts";
 
 function App() {
     const [tasks, setTasks] = useState<ToDoTask[]>([]);
@@ -8,42 +8,40 @@ function App() {
     const [error, setError] = useState<string>("");
 
     useEffect(() => {
-        const fetchTasks = async() => {
-            try{
+        const fetchTasks = async () => {
+            try {
                 const result = await getTasks();
                 setTasks(result);
             }
-            catch(error)
-            {
+            catch (error) {
                 if (error instanceof Error)
                     setError(error.message);
-                else 
+                else
                     setError("Unexpected error");
                 setTasks([]);
             }
-            finally
-            {
+            finally {
                 setIsLoading(false);
             }
         }
-        
+
         void fetchTasks();
     }, []);
-    
-    if (isLoading) 
+
+    if (isLoading)
         return <p>Loading...</p>;
-    
-    if (error) 
+
+    if (error)
         return <p>Error: {error}</p>;
-    
+
     const listTasks = tasks.map(task => <li key={task.id}>{task.title}</li>);
-    
-  return (
-    <>
-        <h1>Task Manager</h1>
-        <ul>{listTasks}</ul>
-    </>
-  )
+
+    return (
+        <>
+            <h1>Task Manager</h1>
+            <ul>{listTasks}</ul>
+        </>
+    )
 }
 
 export default App
