+++ taskmanagementsystem.client/vite.config.ts
@@ -1,0 +1,18 @@
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+
+// https://vite.dev/config/
+export default defineConfig({
+    plugins: [react()],
+    server: {
+        port: 5118,
+        strictPort: true,
+        proxy: {
+            '/api': {
+                target: 'http://localhost:5113',
+                changeOrigin: true,
+                secure: false
+            }
+        }
+    }
+})
